feat: Working with ARK Desktop Wallet export

+++ resources/js/components/ImportModalComponent.vue
@@ -38,27 +38,94 @@
                 this.$store.commit('handleImportModal');
             },
 
+            async getAddressesFromArkWalletJSON(data) {
+                return data.wallets.map(wallet => wallet.address);
+            },
+
+            async makeWalletFromAddress(address, type) {
+                const apiUrl = await this.getApiForType(type);
+                const explorerUrl = await this.getExplorerForType(type);
+                const [walletBalance, delegatePublicKey] = await this.getWalletBalanceAndDelegatePublicKey(apiUrl.url, address);
+
+                const walletDelegate = await this.getDelegateData(apiUrl.url, delegatePublicKey);
+                const delegateShare = await this.getDelegateShare(type, walletDelegate.data.data.username);
+                
+                // Generate a random number for the ID of the Wallet, we add the walletBalance to avoid having the same IDs twice while importing
+                const walletId = Vue.moment().unix() + walletBalance;
+                const arkvatarUrl = await this.getArkvatar(walletDelegate.data.data.address);
+
+                return {
+                    'id': walletId,
+                    'address': address,
+                    'walletBalance': walletBalance,
+                    'arkvatarUrl': arkvatarUrl,
+                    'type': type,
+                    'apiUrl': apiUrl.url,
+                    'explorerUrl': explorerUrl.url,
+                    'delegateUsername': delegateShare.data.name || 'Unknown',
+                    'delegateAddress': walletDelegate.data.data.address || 'Unknown',
+                    'delegatePublicKey': delegatePublicKey || 'Unknown',
+                    'delegateVotesTotal': walletDelegate.data.data.votes || 'Unknown',
+                    'delegateRank': walletDelegate.data.data.rank || 'Unknown',
+                    'delegateSharePercentage': delegateShare.data.payout_percent || 'Unknown',
+                    'delegatePayoutInterval': delegateShare.data.payout_interval || 'Unknown'
+                };
+            },
+
             async submit() {
                 try {
                     let validatedInput = JSON.parse(this.walletsJson);
+                    let existing = localStorage.getItem("addresses");
+                    existing = existing ? JSON.parse(existing) : [];
 
                     if (validatedInput) {
-                        let existing = localStorage.getItem("addresses");
-                        existing = existing ? JSON.parse(existing) : [];
+                        if (validatedInput.hasOwnProperty('meta')) {
+                            const importedAddresses = await this.getAddressesFromArkWalletJSON(validatedInput);
+                            
+                            let type = validatedInput.network.name;
+                            type = type.charAt(0) + type.slice(1).toLowerCase();
+                            
+                            //let temp = [];
 
-                        let merged = existing.concat(validatedInput);
+                            for (let address of importedAddresses) {
+                                let wallet = await this.makeWalletFromAddress(address, type);
+                                this.$store.commit('addWallet', wallet);
+                                //temp.push(wallet);
+                            }
+                            
+                            /*
+                            const waitFor = (ms) => new Promise(r => setTimeout(r, ms))
+                            
+                            await importedAddresses.forEach(async (address) => {
+                                await waitFor(50); 
+                                this.$store.commit('addWallet', await this.makeWalletFromAddress(address, type));
+                                //temp.push(wallet);
+                            });
+                            */
 
-                        let filtered = _.uniqBy(merged, "address");
+                            if(existing.length > 0) {
+                                let merged = existing.concat(this.$store.state.wallets);
+                                let filtered = _.uniqBy(merged, "address");
+                                localStorage.setItem("addresses", JSON.stringify(filtered));
+                            } else {
+                                localStorage.setItem("addresses", JSON.stringify(this.$store.state.wallets));
+                            }
+                        } else {                            
+                            let merged = existing.concat(validatedInput);
 
-                        localStorage.setItem("addresses", JSON.stringify(filtered));
+                            let filtered = _.uniqBy(merged, "address");
+
+                            localStorage.setItem("addresses", JSON.stringify(filtered));
+                        }
 
                         await this.makeToast("Wallets imported !", "check-circle", "success");
-                        
-                        this.$nextTick(async () => {
-                            window.location.reload();
-                        });
+                            
+                        //this.$nextTick(async () => {
+                        //    window.location.reload();
+                        //});
                     }
                 } catch (e) {
+                    console.log(e);
                     console.log("Invalid json");
 
                     await this.makeToast("Invalid JSON format", "times-circle", "error");
